Fixed problem with time overflow at ~4200 seconds

+++ src/platforms/arduino_platform/arduino_systemtime.cpp
@@ -3,13 +3,13 @@
 
 int64_t NOW() {
 
-    static volatile int64_t g_lastMicroseconds = 0;
+    static volatile uint32_t g_lastMicroseconds = 0;
     static volatile int64_t g_currentTime = 0;
 
     uint32_t time = micros();
 
     if (time != g_lastMicroseconds) {
-        g_currentTime = g_currentTime + (int64_t(time) - g_lastMicroseconds)*MICROSECONDS;
+        g_currentTime = g_currentTime + int64_t(time - g_lastMicroseconds)*MICROSECONDS;
         g_lastMicroseconds = time;
     }
